requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
mysql-connector-python==9.1.0
PyJWT>=2.8.0
requests==2.31.0
